ui-requirements.txt
bcrypt==4.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.1
idna==3.11
invoke==2.2.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
paho-mqtt==2.1.0
paramiko==4.0.0
platformdirs==4.4.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.0
requests==2.32.5
rich==14.1.0
scapy==2.6.1
textual==6.1.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.5.0


